refactor pending organizations to use wrapper component
